requirements.txt
requests==2.20.0
zulip
nltk==3.4.5
httplib2==0.10.3
pycountry
newsapi-python
pydictionary
